Travis CI improvements to mirror nengo-loihi
- Simplified `.travis.yml` to use a `MODE` variable to determine
  what script to run, rather than mutually exclusive flags.
- Scripts no longer have `usage` function, but use simple reporting
  if a valid command is not passed (less documentation to update).
- Added `common.sh` with the `exe` function that echoes a command
  and keeps a `STATUS` variable that tracks whether the command
  succeeded.
- Added a number of checks to `static.sh`, including tests on
  notebooks, `gitlint` for commit messages, and `shellcheck` for
  these static scripts.
- Removed `coverage.sh` and integrated it into `test.sh`.

+++ .ci/conda.sh
@@ -1,31 +1,27 @@
 #!/usr/bin/env bash
+if [[ ! -e .ci/common.sh || ! -e nengo ]]; then
+    echo "Run this script from the root directory of this repository"
+    exit 1
+fi
+source .ci/common.sh
+
+# This script sets up the conda environment for all the other scripts
 
 NAME=$0
 COMMAND=$1
 MINICONDA="http://repo.continuum.io/miniconda/Miniconda-latest-Linux-x86_64.sh"
 
-function usage {
-    echo "usage: $NAME <command>"
-    echo
-    echo "  install  Install miniconda and a simple test environment"
-    exit 1
-}
-
 if [[ "$COMMAND" == "install" ]]; then
     wget "$MINICONDA" --quiet -O miniconda.sh
     bash miniconda.sh -b -p "$HOME/miniconda"
     export PATH="$HOME/miniconda/bin:$PATH"
-    conda config --set always_yes yes --set changeps1 no
-    conda update --quiet conda
-    conda info -a
-    conda create --quiet -n test python="$PYTHON" pip
-    source activate test
+    exe conda config --set always_yes yes --set changeps1 no
+    exe conda update --quiet conda
+    exe conda info -a
+    exe conda create --quiet -n test python="$PYTHON" pip
+    exe source activate test
+elif [[ -z "$COMMAND" ]]; then
+    echo "$NAME requires a command like 'install' or 'script'"
 else
-    if [[ -z "$COMMAND" ]]; then
-        echo "Command required"
-    else
-        echo "Command $COMMAND not recognized"
-    fi
-    echo
-    usage
+    echo "$NAME does not define $COMMAND"
 fi
